Fin projet

+++ LITReview/Webapp/views.py
@@ -1,4 +1,3 @@
-from django.forms import CharField
 from django.shortcuts import redirect, render
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
@@ -16,41 +15,47 @@ from itertools import chain
 @login_required
 def index(request):
     user = request.user
-    request_followed = UserFollow.objects.filter(user=user)
-    followed = [user]
-    for people in request_followed:
-        followed.append(people.followed_user)
+    print(type(user))
+    request_followed = User.objects.filter(
+        Q(followed_by__user=user) | Q(pk=user.pk)
+        )
 
-    print(followed)
-    
-    tickets = Ticket.objects.filter(user__in=followed).annotate(type=Value('TICKET', CharField()))
+    tickets = Ticket.objects.filter(
+        user__in=request_followed
+        ).annotate(type=Value('TICKET', CharField()))
+
     reviews = Review.objects.filter(
-                            Q(user__in=followed) | Q(ticket__user=user)
-                            ).annotate(type=Value('REVIEW', CharField()))
+        Q(user__in=request_followed) | Q(ticket__user=user)
+        ).annotate(type=Value('REVIEW', CharField()))
 
     posts = sorted(chain(reviews, tickets),
-                    key=lambda post: post.time_created,
-                    reverse=True)
+                   key=lambda post: post.time_created,
+                   reverse=True)
 
     return render(request,
-            "Webapp/index.html",
-            {"posts": posts})
+                  "Webapp/index.html",
+                  {"posts": posts})
 
 
+@login_required(login_url="/")
 def posts(request):
     user = request.user
 
-    tickets = Ticket.objects.filter(user=user).annotate(type=Value('TICKET', CharField()))
-    reviews = Review.objects.filter(user=user).annotate(type=Value('REVIEW', CharField()))
+    tickets = Ticket.objects.filter(
+        user=user
+        ).annotate(type=Value('TICKET', CharField()))
+
+    reviews = Review.objects.filter(
+        user=user
+        ).annotate(type=Value('REVIEW', CharField()))
 
     posts = sorted(chain(reviews, tickets),
-                    key=lambda post: post.time_created,
-                    reverse=True)
-                    
-    return render(request,
-            "Webapp/posts.html",
-            {"posts" : posts})
+                   key=lambda post: post.time_created,
+                   reverse=True)
 
+    return render(request,
+                  "Webapp/posts.html",
+                  {"posts": posts})
 
 
 def register(request):
@@ -62,23 +67,26 @@ def register(request):
 
     else:
         form = UserCreationForm()
-        
+
     return render(request,
-            "Registration/register.html",
-            {"form": form})
+                  "Registration/register.html",
+                  {"form": form})
+
 
 def registered(request):
     return render(request,
-            "Registration/registered.html",)
+                  "Registration/registered.html",)
 
 
+@login_required(login_url="/")
 def ticket(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
     return render(request,
-                "Webapp/ticket.html",
-                {"ticket": ticket})
+                  "Webapp/ticket.html",
+                  {"ticket": ticket})
 
 
+@login_required(login_url="/")
 def ticket_create(request):
     if request.method == "POST":
         ticket_form = TicketForm(request.POST, request.FILES)
@@ -87,68 +95,81 @@ def ticket_create(request):
             ticket.user = request.user
             ticket.save()
             return redirect("ticket-profil", ticket.id)
-    
+
     else:
         ticket_form = TicketForm()
 
     return render(request,
-            "Webapp/ticket_create.html",
-            {"ticket_form": ticket_form},)
+                  "Webapp/ticket_create.html",
+                  {"ticket_form": ticket_form},)
 
 
+@login_required(login_url="/")
 def ticket_change(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
+
+    if request.user != ticket.user:
+        return redirect("not_allowed")
 
     if request.method == "POST":
         ticket_form = TicketForm(request.POST, request.FILES, instance=ticket)
         if ticket_form.is_valid():
             ticket.save()
             return redirect("ticket-profil", ticket.id)
-    
+
     else:
         ticket_form = TicketForm(instance=ticket)
 
     return render(request,
-                "Webapp/ticket_create.html",
-                {"ticket_form": ticket_form})
+                  "Webapp/ticket_create.html",
+                  {"ticket_form": ticket_form})
 
 
+@login_required(login_url="/")
 def ticket_delete(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
+    if request.user != ticket.user:
+        return redirect("not_allowed")
 
     if request.method == 'POST':
         ticket.delete()
         return redirect('index')
 
     return render(request,
-                "Webapp/ticket_delete.html",
-                {"ticket": ticket})
+                  "Webapp/ticket_delete.html",
+                  {"ticket": ticket})
 
 
+@login_required(login_url="/")
 def review(request, review_id):
     review = Review.objects.get(id=review_id)
     return render(request,
-                "Webapp/review.html",
-                {"review": review})
+                  "Webapp/review.html",
+                  {"review": review})
 
 
+@login_required(login_url="/")
 def review_create(request, ticket_id):
 
     if ticket_id != 0:
         ticket = Ticket.objects.get(id=ticket_id)
-    
+
     else:
         ticket = None
 
     if request.method == "POST":
+
         ticket_form = TicketForm(request.POST, request.FILES, instance=ticket)
         review_form = ReviewForm(request.POST)
 
         if ticket_form.is_valid() & review_form.is_valid():
+            print(ticket)
+            print(ticket.has_response)
             ticket = ticket_form.save(commit=False)
-            if hasattr(ticket, 'user') == False:
+            if hasattr(ticket, 'user') is False:
                 ticket.user = request.user
             ticket.has_response = True
+            print(ticket.has_response)
             ticket.save()
 
             review = review_form.save(commit=False)
@@ -163,32 +184,69 @@ def review_create(request, ticket_id):
         review_form = ReviewForm()
 
     return render(request,
-                "Webapp/review_create.html",
-                {
-                    "ticket": ticket,
-                    "ticket_form": ticket_form,
-                    "review_form": review_form
-                })
+                  "Webapp/review_create.html",
+                  {
+                      "ticket": ticket,
+                      "ticket_form": ticket_form,
+                      "review_form": review_form
+                  })
 
 
-def review_change(request):
-    return
+@login_required(login_url="/")
+def review_change(request, review_id):
+    review = Review.objects.get(id=review_id)
+    ticket = review.ticket
+    if request.user != review.user:
+        return redirect("not_allowed")
+
+    if request.method == "POST":
+        review_form = ReviewForm(request.POST, instance=review)
+        ticket_form = TicketForm(request.POST, request.FILES, instance=ticket)
+        if ticket_form.is_valid() & review_form.is_valid():
+            ticket.save()
+            review.save()
+            return redirect("review-profil", review.id)
+
+    else:
+        review_form = ReviewForm(instance=review)
+        ticket_form = TicketForm(instance=ticket)
+
+    return render(request,
+                  "Webapp/review_create.html",
+                  {
+                      "ticket_form": ticket_form,
+                      "review_form": review_form
+                  })
 
 
-def review_delete(request):
-    return
+@login_required(login_url="/")
+def review_delete(request, review_id):
+    review = Review.objects.get(id=review_id)
+    if request.user != review.user:
+        redirect("not_allowed")
+
+    if request.method == 'POST':
+        review.ticket.has_response = False
+        review.ticket.save()
+        review.delete()
+        return redirect('index')
+
+    return render(request,
+                  "Webapp/review_delete.html",
+                  {"review": review})
 
 
+@login_required(login_url="/")
 def follows(request):
     user = request.user
     follows = UserFollow.objects.all().filter(user=user.id)
     followed = UserFollow.objects.all().filter(followed_user=user.id)
+    error = None
 
     if request.method == "POST":
         to_follow = request.POST.get("username")
         follow = UserFollow()
         follow.user = user
-        error = "aaaa"
 
         try:
             follow.followed_user = User.objects.get(username=to_follow)
@@ -204,23 +262,34 @@ def follows(request):
             error = "Vous suivez déjà cet utilisateur"
 
         except ValueError:
-            error = "Vous ne pouvez pas vous suivre"       
+            error = "Vous ne pouvez pas vous suivre"
 
     return render(request,
-                "Webapp/follows.html",
-                {
-                    "user": user,
-                    "follows": follows,
-                    "followed": followed,
-                    "error": error,
-                })
+                  "Webapp/follows.html",
+                  {
+                      "user": user,
+                      "follows": follows,
+                      "followed": followed,
+                      "error": error,
+                  })
 
 
+@login_required(login_url="/")
 def follow_delete(request, follow_id):
     follow = UserFollow.objects.get(id=follow_id)
+
+    if follow.user != request.user:
+        return redirect("not_allowed")
+
     followed = follow.followed_user
     follow.delete()
 
     return render(request,
-                "Webapp/follow_delete.html",
-                {"followed": followed})
+                  "Webapp/follow_delete.html",
+                  {"followed": followed})
+
+
+@login_required(login_url="/")
+def not_allowed(request):
+    return render(request,
+                  "Webapp/not_allowed.html")
